function to allocate memory for an array of nmemb elements of size bytes each and returns a pointer to the allocated memory.

+++ 0x0C-more_malloc_free/2-calloc.c
@@ -15,7 +15,7 @@ void *_calloc(unsigned int nmemb, unsigned int size)
 	int *ar;
 	int i;
 
-	ar = malloc(nmemb  * size);
+	ar = malloc(nmemb  * (int)size);
 	if (ar == NULL)
 		return (NULL);
 	if (size == 0 || nmemb == 0)
